lakectl cmd split part 2b
