+++ cmd/lakectl/cmd/abuse_commit.go
@@ -26,7 +26,8 @@ var abuseCommitCmd = &cobra.Command{
 		amount := Must(cmd.Flags().GetInt("amount"))
 		gapDuration := Must(cmd.Flags().GetDuration("gap"))
 
-		fmt.Printf("Source branch: %s\n", u.String())
+		fmt.Printf("Source branch: %s\n", u)
+
 		generator := stress.NewGenerator("commit", 1, stress.WithSignalHandlersFor(os.Interrupt, syscall.SIGTERM))
 
 		// generate randomly selected keys as input
