+++ cmd/lakectl/cmd/abuse_link_same_object.go
@@ -24,7 +24,7 @@ var abuseLinkSameObjectCmd = &cobra.Command{
 		parallelism := Must(cmd.Flags().GetInt("parallelism"))
 		key := Must(cmd.Flags().GetString("key"))
 
-		fmt.Printf("Source ref: %s\n", u.String())
+		fmt.Printf("Source ref: %s\n", u)
 		fmt.Printf("Object key: %s\n", key)
 
 		generator := stress.NewGenerator("get-and-link", parallelism, stress.WithSignalHandlersFor(os.Interrupt, syscall.SIGTERM))
