+++ cmd/lakectl/cmd/cherry_pick.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"fmt"
 	"net/http"
 
 	"github.com/spf13/cobra"
@@ -27,7 +28,7 @@ var cherryPick = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		ref := MustParseRefURI("ref", args[0])
 		branch := MustParseBranchURI("branch", args[1])
-		Fmt("Branch: %s\n", branch.String())
+		fmt.Printf("Branch: %s\n", branch)
 
 		if branch.Repository != ref.Repository {
 			Die("Repository mismatch for destination branch and cherry-pick ref", 1)
