+++ cmd/lakectl/cmd/import.go
@@ -91,7 +91,7 @@ var importCmd = &cobra.Command{
 		for {
 			select {
 			case <-sigCtx.Done():
-				Fmt("\nCanceling import\n")
+				fmt.Printf("\nCanceling import\n")
 				resp, err := client.ImportCancelWithResponse(ctx, toURI.Repository, toURI.Ref, &api.ImportCancelParams{Id: importID})
 				DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusNoContent)
 				Die("Import Canceled", 1)
