+++ cmd/lakectl/cmd/refs_restore.go
@@ -2,15 +2,13 @@ package cmd
 
 import (
 	"encoding/json"
+	"fmt"
 	"io"
 	"net/http"
 
 	"github.com/spf13/cobra"
 	"github.com/treeverse/lakefs/pkg/api"
 )
-
-const metadataDumpTemplate = `{{ .Response | json }}
-`
 
 const refsRestoreSuccess = `
 {{ "All references restored successfully!" | green }}
@@ -28,7 +26,7 @@ Since a bare repo is expected, in case of transient failure, delete the reposito
 	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		repoURI := MustParseRepoURI("repository", args[0])
-		Fmt("Repository: %s\n", repoURI.String())
+		fmt.Printf("Repository: %s\n", repoURI)
 		manifestFileName := Must(cmd.Flags().GetString("manifest"))
 		fp := Must(OpenByPath(manifestFileName))
 		defer func() {
@@ -53,30 +51,8 @@ Since a bare repo is expected, in case of transient failure, delete the reposito
 	},
 }
 
-var refsDumpCmd = &cobra.Command{
-	Use:               "refs-dump <repository uri>",
-	Short:             "Dumps refs (branches, commits, tags) to the underlying object store",
-	Hidden:            true,
-	Args:              cobra.ExactArgs(1),
-	ValidArgsFunction: ValidArgsRepository,
-	Run: func(cmd *cobra.Command, args []string) {
-		repoURI := MustParseRepoURI("repository", args[0])
-		client := getClient()
-		resp, err := client.DumpRefsWithResponse(cmd.Context(), repoURI.Repository)
-		DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusCreated)
-		if resp.JSON201 == nil {
-			Die("Bad response from server", 1)
-		}
-
-		Write(metadataDumpTemplate, struct {
-			Response interface{}
-		}{resp.JSON201})
-	},
-}
-
 //nolint:gochecknoinits
 func init() {
-	rootCmd.AddCommand(refsDumpCmd)
 	rootCmd.AddCommand(refsRestoreCmd)
 
 	refsRestoreCmd.Flags().String("manifest", "", "path to a refs manifest json file (as generated by `refs-dump`). Alternatively, use \"-\" to read from stdin")
